requirements.txt
discord.py<=2.0 # 2.0 will have a bunch of breaking changes, we don't want these
sqlalchemy==1.4
python_weather==0.3.5
wikipedia==1.4.0